refactor(ecs): move commands to system dir
