+++ core/CMakeLists.txt
@@ -77,7 +77,7 @@ set(CUBOS_CORE_SOURCE
     "src/cubos/core/ecs/component/manager.cpp"
     "src/cubos/core/ecs/system/system.cpp"
     "src/cubos/core/ecs/system/dispatcher.cpp"
-    "src/cubos/core/ecs/commands.cpp"
+    "src/cubos/core/ecs/system/commands.cpp"
     "src/cubos/core/ecs/blueprint.cpp"
     "src/cubos/core/ecs/world.cpp"
 )
