+++ core/include/cubos/core/ecs/system/system.hpp
@@ -13,10 +13,10 @@
 #include <unordered_set>
 
 #include <cubos/core/ecs/accessors.hpp>
-#include <cubos/core/ecs/commands.hpp>
 #include <cubos/core/ecs/event_reader.hpp>
 #include <cubos/core/ecs/event_writer.hpp>
 #include <cubos/core/ecs/query.hpp>
+#include <cubos/core/ecs/system/commands.hpp>
 #include <cubos/core/ecs/world.hpp>
 
 namespace cubos::core::ecs
